Created custom triggerAction button component
