+++ apps/frontend/src/components/TriggerActionButton/TriggerActionButton.module.css
@@ -1,0 +1,7 @@
+.btn {
+  font-size: 0.7rem;
+  cursor: pointer;
+  width: 50%;
+  margin: auto;
+  margin-bottom: 5px;
+}
